requirements.txt
py-cord[voice]
python-dotenv
yt-dlp
requests
flask